requirements.txt
flask
requests
pyyamls